Refactor broadcast_events into IPC module
