+++ astoria/consumers/astctl/usercode/log.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 import click
 
-from astoria.common.broadcast_event import UsercodeLogBroadcastEvent
+from astoria.common.ipc import UsercodeLogBroadcastEvent
 from astoria.common.mqtt import BroadcastHelper
 from astoria.consumers.astctl.command import Command
