+++ astoria/consumers/astctl/usercode/trigger.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 import click
 
-from astoria.common.broadcast_event import StartButtonBroadcastEvent
+from astoria.common.ipc import StartButtonBroadcastEvent
 from astoria.common.mqtt import BroadcastHelper
 from astoria.consumers.astctl.command import Command
