+++ tests/managers/astprocd/test_usercode_lifecycle.py
@@ -7,8 +7,8 @@ from typing import IO, Any, List, Optional, Tuple, Type
 
 import pytest
 
-from astoria.common.broadcast_event import UsercodeLogBroadcastEvent
 from astoria.common.config import AstoriaConfig
+from astoria.common.ipc import UsercodeLogBroadcastEvent
 from astoria.common.messages.astdiskd import DiskInfo, DiskType, DiskUUID
 from astoria.common.messages.astmetad import Metadata
 from astoria.common.messages.astprocd import CodeStatus
